init SECTION_SYMTAB section

+++ asmToObjFile/elf64/secManager.h
@@ -3,7 +3,7 @@
 
 typedef struct
 {
-    uint16_t size;
+    uint32_t size;
     Section ** allSections;
 } SectionHeader;
 
@@ -11,6 +11,6 @@ SectionHeader * sectionHeaderInit();
 void sectionHeaderFree(SectionHeader *);
 
 Section * sectionHeaderGetByType(SectionHeader *, SectionType);
-uint16_t sectionHeaderGetIndex(SectionHeader *, SectionType);
+uint32_t sectionHeaderGetIndex(SectionHeader *, SectionType);
 Error sectionHeaderAddDataToSection(SectionHeader *, SectionType, char *, uint64_t, uint64_t *);
